+++ app/src/main/java/au/com/shiftyjelly/pocketcasts/PocketCastsApplication.kt
@@ -31,6 +31,7 @@ import au.com.shiftyjelly.pocketcasts.repositories.podcast.EpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.PodcastManager
 import au.com.shiftyjelly.pocketcasts.repositories.podcast.UserEpisodeManager
 import au.com.shiftyjelly.pocketcasts.repositories.shortcuts.DynamicShortcutsSynchronizer
+import au.com.shiftyjelly.pocketcasts.repositories.stats.PlaybackStatsSyncWorker
 import au.com.shiftyjelly.pocketcasts.repositories.support.DatabaseExportHelper
 import au.com.shiftyjelly.pocketcasts.repositories.user.StatsManager
 import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
@@ -301,6 +302,9 @@ class PocketCastsApplication :
         if (FeatureFlag.isEnabled(Feature.SYNC_EOY_DATA_ON_STARTUP)) {
             applicationScope.launch { endOfYearSync.sync() }
         }
+
+        PlaybackStatsSyncWorker.scheduleOneTimeWork(this)
+        PlaybackStatsSyncWorker.schedulePeriodicWork(this)
 
         Timber.i("Launched ${BuildConfig.APPLICATION_ID}")
     }