+++ modules/features/podcasts/src/main/java/au/com/shiftyjelly/pocketcasts/podcasts/view/podcast/PodcastHeader.kt
@@ -99,6 +99,7 @@ import androidx.compose.ui.unit.constrainHeight
 import androidx.compose.ui.unit.constrainWidth
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
+import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import au.com.shiftyjelly.pocketcasts.compose.AppThemeWithBackground
 import au.com.shiftyjelly.pocketcasts.compose.Devices
 import au.com.shiftyjelly.pocketcasts.compose.components.ExpandableText
@@ -114,6 +115,8 @@ import au.com.shiftyjelly.pocketcasts.podcasts.view.components.ratings.PodcastRa
 import au.com.shiftyjelly.pocketcasts.podcasts.viewmodel.PodcastRatingsViewModel.RatingState
 import au.com.shiftyjelly.pocketcasts.repositories.images.PocketCastsImageRequestFactory
 import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
 import coil3.compose.rememberAsyncImagePainter
 import kotlin.math.roundToInt
 import au.com.shiftyjelly.pocketcasts.images.R as IR
@@ -126,6 +129,7 @@ internal fun PodcastHeader(
     title: String,
     category: String,
     author: String,
+    explicit: Boolean,
     description: AnnotatedString,
     podcastInfoState: PodcastInfoState,
     rating: RatingState,
@@ -196,6 +200,7 @@ internal fun PodcastHeader(
                 title = title,
                 category = category,
                 author = author,
+                explicit = explicit,
                 rating = rating,
                 onClickRating = onClickRating,
                 isFollowed = isFollowed,
@@ -234,6 +239,7 @@ private fun PodcastControls(
     title: String,
     category: String,
     author: String,
+    explicit: Boolean,
     rating: RatingState,
     isFollowed: Boolean,
     areNotificationsEnabled: Boolean,
@@ -268,6 +274,7 @@ private fun PodcastControls(
             PodcastCategoriesLabel(
                 category = category,
                 author = author,
+                explicit = explicit,
                 onClickCategory = onClickCategory,
             )
         }
@@ -319,9 +326,11 @@ private fun PodcastControls(
 private fun PodcastCategoriesLabel(
     category: String,
     author: String,
+    explicit: Boolean,
     onClickCategory: () -> Unit,
 ) {
-    val text = remember(category, author, onClickCategory) {
+    val showExplicitIndicator by FeatureFlag.isEnabledFlow(Feature.EXPLICIT_PODCAST_INDICATOR).collectAsStateWithLifecycle()
+    val text = remember(category, author, explicit, onClickCategory, showExplicitIndicator) {
         val text = listOf(category, author).filter(String::isNotBlank).joinToString(separator = " · ")
         buildAnnotatedString {
             append(text)
@@ -339,13 +348,38 @@ private fun PodcastCategoriesLabel(
                     end = category.length,
                 )
             }
+            if (showExplicitIndicator && explicit) {
+                if (text.isNotBlank() || category.isNotBlank()) {
+                    append(" · ")
+                }
+                appendInlineContent("explicit")
+            }
         }
     }
 
+    val iconColor = MaterialTheme.theme.colors.primaryText02
     TextP60(
         text = text,
-        color = MaterialTheme.theme.colors.primaryText02,
+        color = iconColor,
         textAlign = TextAlign.Center,
+        inlineContent = if (showExplicitIndicator && explicit) {
+            mapOf(
+                "explicit" to InlineTextContent(Placeholder(13.sp, 13.sp, PlaceholderVerticalAlign.TextCenter)) {
+                    val explicitDescription = stringResource(LR.string.explicit)
+                    Icon(
+                        painter = painterResource(IR.drawable.explicit),
+                        tint = iconColor,
+                        contentDescription = null,
+                        modifier = Modifier
+                            .fillMaxSize()
+                            // Use this instead of contentDescription so TalkBack only says "Explicit" and doesn't say "Image"
+                            .clearAndSetSemantics { contentDescription = explicitDescription },
+                    )
+                },
+            )
+        } else {
+            emptyMap()
+        },
         modifier = Modifier.padding(bottom = 12.dp),
     )
 }
@@ -934,6 +968,7 @@ private fun PodcastHeaderPreview(
                 title = "The Pitchfork Review",
                 category = "Music",
                 author = "Pitchfork",
+                explicit = true,
                 description = AnnotatedString(
                     """
                     |Savor & Stir is a culinary podcast exploring flavors, techniques, and food stories from chefs and home cooks.