+++ modules/features/profile/src/main/java/au/com/shiftyjelly/pocketcasts/profile/ProfileViewModel.kt
@@ -12,6 +12,7 @@ import au.com.shiftyjelly.pocketcasts.repositories.user.StatsManager
 import au.com.shiftyjelly.pocketcasts.repositories.user.UserManager
 import au.com.shiftyjelly.pocketcasts.utils.Gravatar
 import au.com.shiftyjelly.pocketcasts.utils.toDurationFromNow
+import com.automattic.eventhorizon.BlogsShownEvent
 import com.automattic.eventhorizon.DownloadsShownEvent
 import com.automattic.eventhorizon.EndOfYearProfileCardShownEvent
 import com.automattic.eventhorizon.EndOfYearProfileCardTappedEvent
@@ -186,6 +187,7 @@ class ProfileViewModel @Inject constructor(
             ProfileSection.CloudFiles -> UploadedFilesShownEvent
             ProfileSection.Starred -> StarredShownEvent
             ProfileSection.Bookmarks -> ProfileBookmarksShowEvent
+            ProfileSection.Blogs -> BlogsShownEvent
             ProfileSection.ListeningHistory -> ListeningHistoryShownEvent
             ProfileSection.Help -> SettingsHelpShownEvent
         }