+++ CHANGELOG.md
@@ -1,11 +1,22 @@
+8.12
+-----
+*   Bug Fixes
+    *   Restore episode duration after download for feeds without an RSS length attribute
+        ([#5258](https://github.com/Automattic/pocket-casts-android/pull/5258))
+
 8.11
 -----
+*   New Features
+    *   Added explicit indicator to podcast page
+        ([#5243](https://github.com/Automattic/pocket-casts-android/pull/5243))
 *   Bug Fixes
     *   Fix Skip Last calculation at increased playback speeds
         ([#5230](https://github.com/Automattic/pocket-casts-android/pull/5230))
 *   Updates
     *   Make sure archived episode status is preserved across sync
         ([#5227](https://github.com/Automattic/pocket-casts-android/pull/5227))
+    *   Display dedicated error message for retryable playback errors
+        ([#5244](https://github.com/Automattic/pocket-casts-android/pull/5244))
 
 8.10
 -----